Correct return type of OffsetSegmentString::size() to be size_t
git-svn-id: http://svn.osgeo.org/geos/trunk@3809 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ include/geos/operation/buffer/OffsetSegmentString.h
@@ -171,7 +171,7 @@ public:
 		return ret;
 	}
 
-	inline int size() const { return ptList ? ptList->size() : 0 ; }
+	inline size_t size() const { return ptList ? ptList->size() : 0 ; }
 
 };
